./mtca4u read DUMMY1 "" WORD_CLK_MUX  0 0 raw Causes a segmentation fault
Fixed This bug. New behavior:
./mtca4u read DUMMY1 "" WORD_CLK_MUX  0 0 raw -> returns nothing

Reason:
The readReg method register accessor interface accepts number of bytes. This is
provided as (numElemnts * 4). Providing number of bytes as 0 causes readReg to
return the whole register.

At the same time the read method from the register accessor interface also has
this parameter which takes in the number of elements. For this method the
behavior is diffrent in that it does not return anything when the user has
invoked it with the numElements parameter as 0. (This behavior make more
intiutive sense to me)

The raw mode uses readReg while the normal mode uses read. The inconsiteny in
the behaviour of these lowerlevel methods (in regards to handling numElements)
was a cause for this bug. Due to the inconsistency we end up allocating a vector
with a size that can not accomodate all elemnts that are returned when readReg
is invoked with numElements/numBytes == 0

This might probably be the reason for the segfault seen in issue #1562
mtca4u read_dma FMC25 '' AREA_MULTIPLEXED_SEQUENCE_DAQ 4 0
(the readDMA method from this interface also has the same behavior as readReg)

git-svn-id: https://svnsrv.desy.de/desy/mtca4u/CommandLineTools/trunk@1171 5f601a5c-88b9-47f3-b851-818f6741d00d