Changed MinorVersion to 6         PatchVersion to 1
The code that will be tagged is not necessarily an interface change. Before
these changes it was possible to invoke read_seq as:
mtca4u read_seq DUMMY1 "" DMA "2" <offset> <numElements>
now it is possible to do
mtca4u read_seq DUMMY1 "" DMA "2 1" <offset> <numElements>

This is more of an extension of the interface without breaking the way existing
changes work. For this reason treating these changes as bug fix/patch level


2nd change going in is:
 - modified help text for read_seq indicating that it is possible to accept a
   [\"sequenceList\"] as shown in the above paragraph



git-svn-id: https://svnsrv.desy.de/desy/mtca4u/CommandLineTools/trunk@1368 5f601a5c-88b9-47f3-b851-818f6741d00d